Cmake fixes for Dyninst dependencies (#3)
* Make Dyninst includes a target-specific attribute
This is preferred over just using include_directories.

Dyninst is supposed to export DYNINST_INCLUDE_DIRS, but it actually
exports DYNINST_INCLUDE_DIR (singular). I'll fix this on the Dyninst
side.

* Add Boost configuration
The public headers of SymtabAPI require Boost, but Dyninst doesn't
correctly export the include dependencies it uses, so we have to do it
manually here.

* Remove MSVC check

We don't support Windows.